requirements.txt
Django==5.0.4
icecream==2.1.3
numpy==1.26.4
pandas==2.2.1
pandas-datareader==0.10.0
plotly==5.20.0
yfinance==0.2.37
setuptools==69.2.0
TA_Lib-0.4.28-cp312-cp312-win_amd64